move check_and_translate_hosts into AppProcessesConfig
